requirements.txt
Django>=4.0
djangorestframework>=3.14
django-filter>=23.0
requests>=2.31
beautifulsoup4>=4.12
lxml>=4.9
celery>=5.3
redis>=5.0